--- a/pandas/core/dtypes/concat.py
+++ b/pandas/core/dtypes/concat.py
@@ -7,6 +7,8 @@ from typing import TYPE_CHECKING
 
 import numpy as np
 
+from pandas._libs import lib
+
 from pandas.core.dtypes.astype import astype_array
 from pandas.core.dtypes.cast import (
     common_dtype_categorical_compat,
@@ -29,6 +31,14 @@ if TYPE_CHECKING:
     from pandas.core.arrays import Categorical
 
 
+def _is_nonempty(x, axis) -> bool:
+    # filter empty arrays
+    # 1-d dtypes always are included here
+    if x.ndim <= axis:
+        return True
+    return x.shape[axis] > 0
+
+
 def concat_compat(to_concat, axis: AxisInt = 0, ea_compat_axis: bool = False):
     """
     provide concatenation of an array of arrays each of which is a single
@@ -48,21 +58,13 @@ def concat_compat(to_concat, axis: AxisInt = 0, ea_compat_axis: bool = False):
     -------
     a single array, preserving the combined dtypes
     """
-
-    # filter empty arrays
-    # 1-d dtypes always are included here
-    def is_nonempty(x) -> bool:
-        if x.ndim <= axis:
-            return True
-        return x.shape[axis] > 0
-
     # If all arrays are empty, there's nothing to convert, just short-cut to
     # the concatenation, #3121.
     #
     # Creating an empty array directly is tempting, but the winnings would be
     # marginal given that it would still require shape & dtype calculation and
     # np.concatenate which has them both implemented is compiled.
-    non_empties = [x for x in to_concat if is_nonempty(x)]
+    non_empties = [x for x in to_concat if _is_nonempty(x, axis)]
     if non_empties and axis == 0 and not ea_compat_axis:
         # ea_compat_axis see GH#39574
         to_concat = non_empties
@@ -70,13 +72,13 @@ def concat_compat(to_concat, axis: AxisInt = 0, ea_compat_axis: bool = False):
     dtypes = {obj.dtype for obj in to_concat}
     kinds = {obj.dtype.kind for obj in to_concat}
     contains_datetime = any(
-        isinstance(dtype, (np.dtype, DatetimeTZDtype)) and dtype.kind in ["m", "M"]
+        isinstance(dtype, (np.dtype, DatetimeTZDtype)) and dtype.kind in "mM"
         for dtype in dtypes
     ) or any(isinstance(obj, ABCExtensionArray) and obj.ndim > 1 for obj in to_concat)
 
     all_empty = not len(non_empties)
-    single_dtype = len({x.dtype for x in to_concat}) == 1
-    any_ea = any(isinstance(x.dtype, ExtensionDtype) for x in to_concat)
+    single_dtype = len(dtypes) == 1
+    any_ea = any(isinstance(x, ExtensionDtype) for x in dtypes)
 
     if contains_datetime:
         return _concat_datetime(to_concat, axis=axis)
@@ -308,7 +310,7 @@ def _concat_datetime(to_concat, axis: AxisInt = 0):
 
     to_concat = [ensure_wrapped_if_datetimelike(x) for x in to_concat]
 
-    single_dtype = len({x.dtype for x in to_concat}) == 1
+    single_dtype = lib.dtypes_all_equal([x.dtype for x in to_concat])
 
     # multiple types, need to coerce to object
     if not single_dtype: