--- a/pandas/tests/extension/test_numpy.py
+++ b/pandas/tests/extension/test_numpy.py
@@ -346,18 +346,7 @@ class TestMissing(BaseNumPyTests, base.BaseMissingTests):
 
 
 class TestReshaping(BaseNumPyTests, base.BaseReshapingTests):
-    @pytest.mark.parametrize(
-        "in_frame",
-        [
-            True,
-            pytest.param(
-                False,
-                marks=pytest.mark.xfail(reason="PandasArray inconsistently extracted"),
-            ),
-        ],
-    )
-    def test_concat(self, data, in_frame):
-        super().test_concat(data, in_frame)
+    pass
 
 
 class TestSetitem(BaseNumPyTests, base.BaseSetitemTests):