xref #23362 (closes?  not sure how much more we can do)

```
import numpy as np
import pandas as pd

a = np.random.randint(2000, 2100, size=1000)
b = np.random.randint(2000, 2100, size=1000)

x = pd.core.arrays.period_array(a, freq='B')
y = pd.core.arrays.period_array(b, freq='B')

s = pd.Series(x)
t = pd.Series(y)


In [2]: %timeit pd.concat([s, t], ignore_index=True)
54.6 µs ± 2.41 µs per loop (mean ± std. dev. of 7 runs, 10,000 loops each)  # <- PR
81.5 µs ± 419 ns per loop (mean ± std. dev. of 7 runs, 10,000 loops each)  # <- main
```

cc @phofl do we need to do anything else refs-wise? when calling Manager.from_array?