@phofl 2023-03-30
I think we only get there when we have ``axis=0`` which creates a copy anyway, so no we are good